Created main.py and text file
In main.py, created a subroutine to read a file, in this case calculator_options.txt. In calculator_options.txt created a menu for calculator options.
In main.py created a subroutine and while statement that allows a user to select an option from the menu provided to do calculations using two user input numbers