Created NumberGuessingGame folder
Inside the folder created main.py that includes a function that generates a random number between 1 and 10. Another function that uses a while statement that allows user to try and guess the correct number. If they run out of attempts or correctly guess the number they have the option to play again